MNT: Move mixins into separate files to isolate dependencies.
Includes module docstrings with working examples

Also, minor changes:
- renamed new_uid -> new_short_uid
- added new_uid, provides untrucated str(uuid.uuid4)
- new_short_uid is for filenames, new_uid for datum ids
